updates product controller and routes

+++ server/controllers/productController.js
@@ -1,6 +1,7 @@
 import mongoose from "mongoose";
 import Product from "../models/productModel.js";
 
+// controller to get all products
 export const getProducts = async (req, res) => {
   try {
     const products = await Product.find({});
@@ -11,6 +12,7 @@ export const getProducts = async (req, res) => {
   }
 };
 
+// controller to create product
 export const createProduct = async (req, res) => {
   const product = req.body; // user will send this data
 
@@ -42,9 +44,16 @@ export const createProduct = async (req, res) => {
   }
 };
 
+// controller to update product
 export const updateProduct = async (req, res) => {
   const { id } = req.params;
   const product = req.body; // user will be able to update the data
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid Product Id" });
+  }
 
   try {
     const updatedProduct = await Product.findByIdAndUpdate(id, product, {
@@ -63,14 +72,22 @@ export const updateProduct = async (req, res) => {
   }
 };
 
+// controller to delete product
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
-  console.log("id: ", id);
+  //   console.log("id: ", id);
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid Product Id" });
+  }
+
   try {
     await Product.findByIdAndDelete(id); // delete the product
     res.status(200).json({ success: true, message: "Product deleted" }); // send response
   } catch (err) {
     console.log("Error while deleting product", err);
-    res.status(404).json({ success: false, message: "Product not found" }); // send response
+    res.status(500).json({ success: false, message: "Internal Server Error" }); // send response
   }
 };
